% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{density_mixt_normal}
\alias{density_mixt_normal}
\title{probability density function of mixtured normal distribution}
\usage{
density_mixt_normal(x, mu, sigma, pi)
}
\arguments{
\item{x}{sample point}

\item{mu}{population means}

\item{sigma}{population sd}

\item{pi}{mixtured rate}
}
\description{
probability density function of mixtured normal distribution
}
