% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{LL_mixt_normal}
\alias{LL_mixt_normal}
\title{LL function : log likelihood of mixtured normal distribution}
\usage{
LL_mixt_normal(x, mu, sigma, pi)
}
\arguments{
\item{x}{sample point}

\item{mu}{population means}

\item{sigma}{population sd}

\item{pi}{mixtured rate}
}
\description{
LL function : log likelihood of mixtured normal distribution
}
