% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{plot.EM_MixtNormal}
\alias{plot.EM_MixtNormal}
\title{Plot of EM_Mixt_Normal}
\usage{
\method{plot}{EM_MixtNormal}(x, ...)
}
\arguments{
\item{x}{<EM_MixtNormal> the result object of the function :}

\item{...}{Arguments to be passed to methods.}
}
\description{
Plot of EM_Mixt_Normal
}
