% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{plot_components}
\alias{plot_components}
\title{Plot of histogram of components}
\usage{
plot_components(x)
}
\arguments{
\item{x}{<EM_MixtNormal> the result object of the function :}
}
\description{
Plot of histogram of components
}
