% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{plot_history}
\alias{plot_history}
\title{plotting history of EM-algoritm with gif file}
\usage{
plot_history(result, width = 5, height = 5,
  file_name = "history_of_EM")
}
\arguments{
\item{result}{<EM_MixtNormal>}

\item{width}{<double_scalar> image size}

\item{height}{<double_scalar> image size}

\item{file_name}{<string> gif file name}
}
\description{
plotting history of EM-algoritm with gif file
}
