% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{plot_LL}
\alias{plot_LL}
\title{Plot of history of loglikelihood}
\usage{
plot_LL(x)
}
\arguments{
\item{x}{<EM_MixtNormal> the result object of the function :}
}
\description{
Plot of history of loglikelihood
}
