% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{predict.EM_MixtNormal}
\alias{predict.EM_MixtNormal}
\title{predict component that each datapoint belongs to}
\usage{
\method{predict}{EM_MixtNormal}(object, x, ...)
}
\arguments{
\item{object}{<EM_MixtNormal> result of EM_mixt_normal}

\item{x}{<double_scalar> 1-dim. data points}

\item{...}{additional arguments}
}
\description{
predict component that each datapoint belongs to
}
