% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{print.EM_MixtNormal}
\alias{print.EM_MixtNormal}
\title{print result of EM_Mixt_Normal}
\usage{
\method{print}{EM_MixtNormal}(result)
}
\arguments{
\item{result}{<EM_MixtNormal> the result object of the function : em_mixt_normal}
}
\description{
print result of EM_Mixt_Normal
}
