% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{random_mixt_normal}
\alias{random_mixt_normal}
\title{random sampling of mixtured normal distribution}
\usage{
random_mixt_normal(n, mu, sigma, pi)
}
\arguments{
\item{n}{sample size}

\item{mu}{population means}

\item{sigma}{population sd}

\item{pi}{mixtured rate}
}
\description{
random sampling of mixtured normal distribution
}
