% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{summary.EM_MixtNormal}
\alias{summary.EM_MixtNormal}
\title{summary result of EM_Mixt_Normal}
\usage{
\method{summary}{EM_MixtNormal}(result)
}
\arguments{
\item{result}{<EM_MixtNormal> the result object of the function :}
}
\description{
summary result of EM_Mixt_Normal
}
