% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{density_mixt_normal}
\alias{density_mixt_normal}
\title{Probability Density Function of 1-dim. Mixtures of Normal Distribution}
\usage{
density_mixt_normal(x, mu, sigma, ratio)
}
\arguments{
\item{x}{<double vector> : sample point}

\item{mu}{<double vector> : population means}

\item{sigma}{<double vector> : population sd}

\item{ratio}{<double vector> : mixtured ratio}
}
\value{
The output is <data_frame> that is probability densities of each saple point at each component and mixture distribution.
}
\description{
Set sample point x and parameters of MixtNormal(mu,sigma,pi).
By definition, dimension of mu, sigma and pi must be the same.
In the function, parameter `pi` is normalized.
}
