% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{EM_mixt_normal}
\alias{EM_mixt_normal}
\title{EM Algorithm for 1-dim. Mixtures of Normal Distribution}
\usage{
EM_mixt_normal(x, max_iter, tol, init_mu, init_sigma, init_ratio)
}
\arguments{
\item{x}{<double vector> : sample}

\item{max_iter}{<int scalar> : maximum number of iterations}

\item{tol}{<double scalar> : tolerance}

\item{init_mu}{<double vector> : initial values of population means}

\item{init_sigma}{<double vector> : initial values of population sds}

\item{init_ratio}{<double vector> : initial values of population ratio}
}
\value{
The output is the set (MLE of params, LL, estimated component of each sample point, n_iter).
}
\description{
EM Algorithm for 1-dim. Mixtures of Normal Distribution
}
