% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{EM_mixt_normal}
\alias{EM_mixt_normal}
\title{EM function : EM algorithm for mixtured normal distribution}
\usage{
EM_mixt_normal(x, max_iter, tol, init_mu, init_sigma, init_pi)
}
\arguments{
\item{x}{sample}

\item{max_iter}{maximum number of iterations}

\item{tol}{tolerance}

\item{init_mu}{initial values of population means}

\item{init_sigma}{initial values of population sds}

\item{init_pi}{initial values of population rate}
}
\description{
EM function : EM algorithm for mixtured normal distribution
}
