% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{LL_mixt_normal}
\alias{LL_mixt_normal}
\title{Log Likelihood of 1-dim. Mixtures of Normal Distribution}
\usage{
LL_mixt_normal(x, mu, sigma, pi)
}
\arguments{
\item{x}{<double vector> : sample point}

\item{mu}{<double vector> : population means}

\item{sigma}{<double vector> : population sd}

\item{pi}{<double vector> : mixtured rate}
}
\value{
The output is <double vector> that is loglikelihood of each point of x
}
\description{
Log Likelihood of 1-dim. Mixtures of Normal Distribution
}
