% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{random_mixt_normal}
\alias{random_mixt_normal}
\title{Random Sampling of 1-dim. Mixtures of Normal Distribution}
\usage{
random_mixt_normal(n, mu, sigma, ratio)
}
\arguments{
\item{n}{<int scalar> : sample size}

\item{mu}{<double vector> : population means}

\item{sigma}{<double vector> : population sds}

\item{ratio}{<double vector> : mixtured ratio}
}
\value{
The output will be <double vector> of size n that you input.
}
\description{
Set sample size n and parameters of MixtNormal(mu,sigma,pi).
By definition, dimension of mu, sigma and pi must be the same.
In the function, parameter `pi` is normalized.
}
