% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt_normal.R
\name{summary.EM_MixtNormal}
\alias{summary.EM_MixtNormal}
\title{summary result of EM_Mixt_Normal}
\usage{
\method{summary}{EM_MixtNormal}(object, ...)
}
\arguments{
\item{object}{<EM_MixtNormal> the result object of the function :}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
summary.EM_MixtNormal used to procedure result summaries of the result of the function : EM_mixt_normal.
}
